feat: refactor InterviewPage to use server-side data fetching and separate client logic into InterviewPageClient component
